Crate sync15_traits

Re-exports

pub use request::CollectionRequest;

pub use request::RequestOrder;

Modules

bridged_engine

changeset

client
This module has to be here because of some hard-to-avoid hacks done for the tabs engine... See issue #2590

payload

request

server_timestamp

store

telemetry
Manage recording sync telemetry. Assumes some external telemetry library/code which manages submitting.

Structs

ApplyResults

CollSyncIds

Guid
This is a type intended to be used to represent the guids used by sync. It has several benefits over using a String:

Payload
Represents the decrypted payload in a Bso. Provides a minimal layer of type safety to avoid double-encrypting.

RecordChangeset

ServerTimestamp
Typesafe way to manage server timestamps without accidentally mixing them up with local ones.

Enums

StoreSyncAssociation
Defines how a store is associated with Sync.

Traits

BridgedEngine
A bridged Sync engine implements all the methods needed to support Desktop Sync.

Store
Low-level store functionality. Stores that need custom reconciliation logic should use this.

Functions

skip_if_default

Type Definitions

IncomingChangeset

OutgoingChangeset