Struct sync15_traits::Payload

pub struct Payload {
    pub id: Guid,
    pub deleted: bool,
    pub data: Map<String, JsonValue>,
}

Represents the decrypted payload in a Bso. Provides a minimal layer of type safety to avoid double-encrypting.

Fields
id: Guiddeleted: booldata: Map<String, JsonValue>
Implementations

impl Payload

pub fn new_tombstone(id: impl Into<Guid>) -> Payload

pub fn new_tombstone_with_ttl(id: impl Into<Guid>, ttl: u32) -> Payload

pub fn with_sortindex(self, index: i32) -> Payload

pub fn with_auto_field<T: Into<JsonValue>>(
self,
name: &str,
v: Option<T>
) -> Payload

“Auto” fields are fields like ‘sortindex’ and ‘ttl’, which are:

Added to the payload automatically when deserializing if present on the incoming BSO or envelope.

Removed from the payload automatically and attached to the BSO or envelope if present on the outgoing payload.

pub fn take_auto_field<V>(&mut self, name: &str) -> Option<V> where
V: Deserialize<'a>,

pub fn id(&self) -> &str

pub fn is_tombstone(&self) -> bool

pub fn from_json(value: JsonValue) -> Result<Payload, Error>

pub fn into_record<T>(self) -> Result<T, Error> where
T: Deserialize<'a>,

Deserializes the BSO payload into a specific record type T.

BSO payloads are unstructured JSON objects, with string keys and dynamically-typed values. into_record makes it more convenient to work with payloads by converting them into data type-specific structs. Your record type only needs to derive or implement serde::Deserialize; Serde will take care of the rest.

Errors

into_record returns errors for type mismatches. As an example, trying to deserialize a string value from the payload into an integer field in T will fail.

If there’s a chance that a field contains invalid or mistyped data, you’ll want to extract it from payload.data manually, instead of using into_record. This has been seen in the wild: for example, dateAdded for bookmarks can be either an integer or a string.

pub fn from_record<T: Serialize>(v: T) -> Result<Payload, Error>

pub fn into_json_string(self) -> String