Trait sync15_traits::SyncEngine

pub trait SyncEngine {
    fn collection_name(&self) -> Cow<'static, str>;

    fn apply_incoming(
        &self, 
        inbound: Vec<IncomingChangeset>, 
        telem: &mut Engine
    ) -> Result<OutgoingChangeset>;

    fn sync_finished(
        &self, 
        new_timestamp: ServerTimestamp, 
        records_synced: Vec<Guid>
    ) -> Result<()>;

    fn get_collection_requests(
        &self, 
        server_timestamp: ServerTimestamp
    ) -> Result<Vec<CollectionRequest>>;

    fn get_sync_assoc(&self) -> Result<EngineSyncAssociation>;

    fn reset(&self, assoc: &EngineSyncAssociation) -> Result<()>;

    fn wipe(&self) -> Result<()>;

    fn prepare_for_sync(
        &self, 
        _get_client_data: &dyn Fn() -> ClientData
    ) -> Result<()> { ... }

    fn set_local_encryption_key(&mut self, _key: &str) -> Result<()> { ... }
}

A “sync engine” is a thing that knows how to sync. It’s often implemented by a “store” (which is the generic term responsible for all storage associated with a component, including storage required for sync.)

Low-level engine functionality. Engines that need custom reconciliation logic should use this.

Different engines will produce errors of different types. To accommodate this, we force them all to return anyhow::Error.

Required methods

fn collection_name(&self) -> Cow<'static, str>

fn apply_incoming(
&self,
inbound: Vec<IncomingChangeset>,
telem: &mut Engine
) -> Result<OutgoingChangeset>

inbound is a vector to support the case where get_collection_requests returned multiple requests. The changesets are in the same order as the requests were – e.g. if vec![req_a, req_b] was returned from get_collection_requests, inbound will have the results from req_a as its first index, and those from req_b as it’s second.

fn sync_finished(
&self,
new_timestamp: ServerTimestamp,
records_synced: Vec<Guid>
) -> Result<()>

fn get_collection_requests(
&self,
server_timestamp: ServerTimestamp
) -> Result<Vec<CollectionRequest>>

The engine is responsible for building the collection request. Engines typically will store a lastModified timestamp and use that to build a request saying “give me full records since that date” - however, other engines might do something fancier. This could even later be extended to handle “backfills” etc

To support more advanced use cases, multiple requests can be returned here - either from the same or different collections. The vast majority of engines will just want to return zero or one item in their vector (zero is a valid optimization when the server timestamp is the same as the engine last saw, one when it is not)

Important: In the case when more than one collection is requested, it’s assumed the last one is the “canonical” one. (That is, it must be for “this” collection, its timestamp is used to represent the sync, etc). (Note that multiple collection request support is currently unused, so it might make sense to delete it - if we need it later, we may find a better shape for our use-case)

fn get_sync_assoc(&self) -> Result<EngineSyncAssociation>

Get persisted sync IDs. If they don’t match the global state we’ll be reset() with the new IDs.

fn reset(&self, assoc: &EngineSyncAssociation) -> Result<()>

Reset the engine (and associated store) without wiping local data, ready for a “first sync”. assoc defines how this store is to be associated with sync.

fn wipe(&self) -> Result<()>

Loading content...
Provided methods

fn prepare_for_sync(
&self,
_get_client_data: &dyn Fn() -> ClientData
) -> Result<()>

Prepares the engine for syncing. The tabs engine currently uses this to store the current list of clients, which it uses to look up device names and types.

Note that this method is only called by sync_multiple, and only if a command processor is registered. In particular, prepare_for_sync will not be called if the store is synced using sync::synchronize or sync_multiple::sync_multiple. It will be called if the store is synced via the Sync Manager.

TODO(issue #2590): This is pretty cludgey and will be hard to extend for any case other than the tabs case. We should find another way to support tabs…

fn set_local_encryption_key(&mut self, _key: &str) -> Result<()>

Tells the engine what the local encryption key is for the data managed by the engine. This is only used by collections that store data encrypted locally and is unrelated to the encryption used by Sync. The intent is that for such collections, this key can be used to decrypt local data before it is re-encrypted by Sync and sent to the storage servers, and similarly, data from the storage servers will be decrypted by Sync, then encrypted by the local encryption key before being added to the local database.

The expectation is that the key value is being maintained by the embedding application in some secure way suitable for the environment in which the app is running - eg, the OS “keychain”. The value of the key is implementation dependent - it is expected that the engine and embedding application already have some external agreement about how to generate keys and in what form they are exchanged. Finally, there’s an assumption that sync engines are short-lived and only live for a single sync - this means that sync doesn’t hold on to the key for an extended period.

This will panic if called by an engine that doesn’t have explicit support for local encryption keys as that implies a degree of confusion which shouldn’t be possible to ignore.

Loading content...