Module sync15_traits::bridged_engine

Structs

ApplyResults

IncomingEnvelope

An envelope for an incoming item, passed to BridgedEngine::store_incoming. Envelopes are a halfway point between BSOs, the format used for all items on the Sync server, and records, which are specific to each engine.

OutgoingEnvelope

An envelope for an outgoing item, returned from BridgedEngine::apply. This is conceptually identical to IncomingEnvelope, but omits fields that are only set by the server, like modified.

Enums

PayloadError

An error that indicates a payload is invalid.

Traits

BridgedEngine

A BridgedEngine acts as a bridge between application-services, rust implemented sync engines and sync engines as defined by Desktop Firefox.