Module sync15_traits::client

This module has to be here because of some hard-to-avoid hacks done for the tabs engine… See issue #2590

Structs

ClientData
Argument to Store::prepare_for_sync. See comment there for more info. Only really intended to be used by tabs engine.

RemoteClient
Information about a remote client in the clients collection.

Enums

DeviceType
The type of a client. Please keep these variants in sync with the device types in the FxA client, tabs and sync manager.